move duplicated code from BeaconService into common class
